backend wqeek2/app/extensions.py
from flask_marshmallow import Marshmallow
from flask_sqlalchemy import SQLAlchemy
from sqlalchemy.orm import DeclarativeBase

class Base(DeclarativeBase):
    pass

ma = Marshmallow()
db = SQLAlchemy(model_class=Base)