backend wqeek2/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
mysql-connector-python==8.2.0
requests==2.31.0
